refactor: Team User Profile -> GP User Profile
