+++ gameplan/gameplan/doctype/gp_user_profile/gp_user_profile.js
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, Frappe Technologies Pvt Ltd and contributors
 // For license information, please see license.txt
 
-frappe.ui.form.on('Team User Profile', {
+frappe.ui.form.on('GP User Profile', {
 	// refresh: function(frm) {
 
 	// }
