+++ gameplan/gameplan/doctype/gp_user_profile/patches/setup_rembg.py
