+++ gameplan/gameplan/doctype/gp_user_profile/profile_photo.py
