+++ gameplan/gameplan/doctype/gp_user_profile/test_gp_user_profile.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestTeamUserProfile(FrappeTestCase):
+class TestGPUserProfile(FrappeTestCase):
 	pass
