+++ gameplan/hooks.py
@@ -125,12 +125,12 @@ doc_events = {
 		"on_trash": "gameplan.mixins.on_delete.on_trash",
 	},
 	"User": {
-		"after_insert": "gameplan.gameplan.doctype.team_user_profile.team_user_profile.create_user_profile",
+		"after_insert": "gameplan.gameplan.doctype.gp_user_profile.team_user_profile.create_user_profile",
 		"on_trash": [
-			"gameplan.gameplan.doctype.team_user_profile.team_user_profile.delete_user_profile",
+			"gameplan.gameplan.doctype.gp_user_profile.team_user_profile.delete_user_profile",
 			"gameplan.gameplan.doctype.gp_guest_access.gp_guest_access.on_user_delete",
 		],
-		"on_update": "gameplan.gameplan.doctype.team_user_profile.team_user_profile.on_user_update"
+		"on_update": "gameplan.gameplan.doctype.gp_user_profile.team_user_profile.on_user_update"
 	}
 }
